+++ src/components/generic/editable-field-addables/index.tsx
@@ -1,5 +1,5 @@
 export { default as MetaformTextFieldAddable } from "./text-field-addable";
-export { default as MetaformHtmlAddable } from "./html-field-addables";
+export { default as MetaformHtmlAddable } from "./html-field-addable";
 export { default as MetaformRadioFieldAddable } from "./radio-field-addable";
-export { default as MetaformSubmitFieldComponent } from "./metaform-submit-field-component";
+export { default as MetaformSubmitFieldAddable } from "./submit-field-addable";
 export { default as MetaformNumberFieldAddable } from "./number-field-addable";
