+++ src/components/generic/editable-field-addables/text-field-addable.tsx
@@ -1,8 +1,6 @@
 import { FormControl, FormLabel, TextField, WithStyles, withStyles } from "@material-ui/core";
 import React from "react";
-import { MetaformField } from "../../../generated/client";
 import strings from "../../../localization/strings";
-import { FieldValue } from "../../../types";
 import styles from "../../../styles/generics/editable-field-components/metaform-text-field-component"
 
 /**
